prevent negative length memmove copy
Discovered using a dcmi power reading command that attempts to return an error 'DCMI request failed because: Request data truncated (c6)'. The function lanplus_crypt.c lanplus_decrypt_payload will decrypt a shorter data length of 7 rather than the expected length of 8 bytes, causing ipmi_lan_poll_single() to incorrectly assume the extra_data_length of -1 should trigger a memmove of rsp->data to rsp->data + offset. This results in a segv in the memmove library. The correction is to test for a positive size of extra_data_length.